refactor: centralize auth verification
