+++ app/api/stats/route.ts
@@ -4,54 +4,16 @@ export const dynamic = "force-dynamic"
 import { type NextRequest, NextResponse } from "next/server"
 import { db } from "@/lib/database"
 import { logger } from "@/lib/logger"
-import jwt from "jsonwebtoken"
-
-const JWT_SECRET = process.env.JWT_SECRET || "fallback-secret-key"
-
-async function verifyAuth(request: NextRequest) {
-  try {
-    // محاولة قراءة التوكن من الكوكيز أولاً
-    let token = request.cookies.get("auth-token")?.value
-
-    // إذا لم يوجد في الكوكيز، جرب Authorization header
-    if (!token) {
-      const authHeader = request.headers.get("authorization")
-      if (authHeader && authHeader.startsWith("Bearer ")) {
-        token = authHeader.split(" ")[1]
-      }
-    }
-
-    if (!token) {
-      logger.warn("No token found in cookies or headers")
-      return { user: null, error: "No token provided" }
-    }
-
-    logger.info("Token found, verifying...")
-
-    // التحقق من صحة التوكن
-    const decoded = jwt.verify(token, JWT_SECRET) as any
-
-    if (!decoded || !decoded.userId) {
-      logger.warn("Invalid token structure:", decoded)
-      return { user: null, error: "Invalid token" }
-    }
-
-    logger.info("Token verified successfully for user:", decoded.userId)
-    return { user: decoded, error: null }
-  } catch (error) {
-    logger.error("Auth verification failed:", error)
-    return { user: null, error: "Authentication failed" }
-  }
-}
+import { verifyAuth } from "@/lib/auth"
 
 export async function GET(request: NextRequest) {
   try {
     logger.info("🔍 GET /api/stats - Starting request")
 
     // التحقق من المصادقة
-    const { user, error } = await verifyAuth(request)
-    if (!user) {
-      logger.warn("❌ Authentication failed:", error)
+    const authResult = await verifyAuth(request)
+    if (!authResult.success || !authResult.user) {
+      logger.warn("❌ Authentication failed:", authResult.message)
       return NextResponse.json({ success: false, error: "غير مصرح" }, { status: 401 })
     }
